HPLT - 22-DEC-2025. Class, Nature & Notes Table style fix

+++ AppliedAccounts/Models/COAClassModel.cs
@@ -63,7 +63,7 @@ namespace AppliedAccounts.Models
         {
             COAClassRecord _Record = new();
             {
-                _Record.ID = (int)_Row["ID"];
+                _Record.ID = (long)_Row["ID"];
                 _Record.Code = (string)_Row["Code"];
                 _Record.Title = (string)_Row["Title"];
 
@@ -204,7 +204,7 @@ namespace AppliedAccounts.Models
 
     public class COAClassRecord
     {
-        public int ID { get; set; }
+        public long ID { get; set; }
         public string Code { get; set; } = string.Empty;
         public string Title { get; set; } = string.Empty;
     }
